Simplify widget state management
Instead of build and update there's now just layout. There's also a new
WidgetState trait that contains a new method and the min_size and
extra_layers that were previously in the Widget trait. The way the state
management was previously, where state is only constructed in layout,
hassome promise, but a lot of problems would need to be solved, because
as it was before there was a lot of unwrapping involved. For the approach
to work it the build method would have to return a type that contains
both the state and the widget and that can then be made into only the
state. Other methods would then be called on this value. This way
variants are sure to match on the type level. But that approach makes
lists more complicated, because this would need to work for whole lists.